Be consistent about allowing regions to have zero length
The text said that a text region consisted of one or more
characters, but later it said that if the end of the region
was not specified, its length was assumed to be 0.

Similarly it said that a binary region consisted of one or
more bytes, but later it said that if the end of the region
was not specified, its length was assumed to be 0.

Change text to say that a text region consists of zero or
more characters, and a binary region consists of zero or
more bytes.