Rename "issue" to "result" throughout
This is more than just a mechanical change. Until now, we have used
the term "issue" to mean both "a condition detected by a tool" and
"a potentially undesirable condition". Not all conditions detected
by tools are undersirable, and so we now use two terms:

* A "result" is any condition detected and reported by a tool.

* A "problem" is a potentially undesirable condition indicated by a result.

In contexts where a result might or might not indicate a problem,
we write "the problem indicated by the result, if any."

We write "detects an issue" to "produces a result", and similar
constructions.

We speak of a rule being "evaluated to produce a result" rather than
being "violated".

We make corresponding changes from "issue log viewer" to "result log
viewer", "issue management system" to "result management system", etc.