Fix #231: Is there a reason to recommend upper-casing?
The original question that prompted this issue was whether it mattered whether tools normalized path names to upper case or to lower case (on case-insensitive but case-preserving file systems). As a result of the discussion in #231, we've agreed not to recommend any case normalization at all.

So we remove the section on case normalization, and all other references to it in Annex E.

@michaelcfanning @nguerrera