Fix #208: Specify numeric and string constraints
We reviewed the string length and numeric value constraints on all SARIF
properties, as well as the length constraints on all array properties,
and the constraints on the number of sub-properties in properties
such as `run.files`.

The following arrays were required to be non-empty if present, and
I did not change them:

* `sarifLog.runs`
* `result.locations`
* `result.stacks`
* `result.relatedLocations`
* `codeFlow.locations`
* `stack.Frames`

`result.CodeFlows` was not required to be non-empty, so I changed it for
consistency with `result.stacks`.

The following arrays that are allowed to be empty:
* `run.results` (a special case)
* `result.fixes` (why?)
* `exception.innerException`

As the text now stands, empty objects ( {} ) are optional, but most
empty arrays are forbidden. This doesn't make sense to me. Let's discuss.

Also:
* Adjusted several places where we expressed the same thing using
  slightly different words in different places. In all such cases,
  I unified on the construct that was used most often.